refactor: split the compose merge tests into their own file
The new union/merge cases pushed internal/compose/extends/merge.rs to
587 code lines, well over the 500-line hard limit, and line-limit.yml
failed the build.

Moved the test module to extends/merge/tests.rs, the same shape the
quadlet and autostart modules already use. merge.rs drops to 275 lines
and the tests file is 311. No test bodies were touched.

Part of #1078
